+++ src/Jellyfin.Plugin.LiveTvGroups/PluginServiceRegistrator.cs
@@ -32,6 +32,7 @@ public class PluginServiceRegistrator : IPluginServiceRegistrator
         serviceCollection.AddHostedService<WebInjectionService>();
         serviceCollection.AddHostedService<PluginPagesService>();
         serviceCollection.AddHostedService<PlaylistSyncTrigger>();
+        serviceCollection.AddSingleton<PlayerService>();
         serviceCollection.AddSingleton<GroupGuideService>();
         serviceCollection.AddSingleton<LiveTvStreamBridge>();
         serviceCollection.AddSingleton<AppGuideService>();